Adding Query Key Authentication for API

+++ actions/user.actions.js
@@ -23,7 +23,7 @@ consoleOutput = (route, request, status) => {
 };
 
 exports.get = async (req, res) => {
-  if (req.query.key === process.env.PASS) {
+  if (req.query.key === process.env.QUERYKEY) {
     const exists = await User.find({});
     if (exists && exists.length > 0) {
       consoleOutput("/api/users", "GET", 200);
@@ -38,44 +38,64 @@ exports.get = async (req, res) => {
 };
 
 exports.getOne = async (req, res) => {
-  const exists = await User.findById(req.params.id);
-  if (exists) {
-    consoleOutput(`/api/users/${req.params.id}`, "GET", 200);
-    return res.status(200).send(exists);
+  if (req.query.key === process.env.QUERYKEY) {
+    const exists = await User.findById(req.params.id);
+    if (exists) {
+      consoleOutput(`/api/users/${req.params.id}`, "GET", 200);
+      return res.status(200).send(exists);
+    }
+    consoleOutput(`/api/users/${req.params.id}`, "GET", 400);
+    return res.status(404).send(error[404]);
+  } else {
+    consoleOutput("/api/users", "GET", 403);
+    return res.status(403).send(error[403]);
   }
-  consoleOutput(`/api/users/${req.params.id}`, "GET", 400);
-  return res.status(404).send(error[404]);
 };
 
 exports.post = async (req, res) => {
-  const exists = await User.findOne(req.body);
-  if (!exists || exists.length === 0) {
-    const user = await new User(req.body).save();
-    consoleOutput("/api/users", "POST", 200);
-    return res.status(200).send(user);
+  if (req.query.key === process.env.QUERYKEY) {
+    const exists = await User.findOne(req.body);
+    if (!exists || exists.length === 0) {
+      const user = await new User(req.body).save();
+      consoleOutput("/api/users", "POST", 200);
+      return res.status(200).send(user);
+    }
+    consoleOutput("/api/users", "POST", 400);
+    return res.status(400).send(error[400]);
+  } else {
+    consoleOutput("/api/users", "GET", 403);
+    return res.status(403).send(error[403]);
   }
-  consoleOutput("/api/users", "POST", 400);
-  return res.status(400).send(error[400]);
 };
 
 exports.put = async (req, res) => {
-  const exists = await User.findById(req.params.id);
-  if (exists) {
-    const update = await User.findByIdAndUpdate(exists._id, req.body);
-    consoleOutput(`/api/users/${req.params.id}`, "PUT", 200);
-    return res.status(200).send(update);
+  if (req.query.key === proccess.env.QUERYKEY) {
+    const exists = await User.findById(req.params.id);
+    if (exists) {
+      const update = await User.findByIdAndUpdate(exists._id, req.body);
+      consoleOutput(`/api/users/${req.params.id}`, "PUT", 200);
+      return res.status(200).send(update);
+    }
+    consoleOutput(`/api/users${req.params.id}`, "PUT", 404);
+    return res.status(404).send(error[404]);
+  } else {
+    consoleOutput("/api/users", "GET", 403);
+    return res.status(403).send(error[403]);
   }
-  consoleOutput(`/api/users${req.params.id}`, "PUT", 404);
-  return res.status(404).send(error[404]);
 };
 
 exports.delete = async (req, res) => {
-  const exists = await User.findById(req.params.id);
-  if (exists) {
-    const response = await User.findByIdAndDelete(exists._id);
-    consoleOutput(`/api/users/${req.params.id}`, "DELETE", 200);
-    return res.status(200).send({ message: `User ${req.params.id} deleted` });
+  if (req.query.key === process.env.QUERYKEY) {
+    const exists = await User.findById(req.params.id);
+    if (exists) {
+      const response = await User.findByIdAndDelete(exists._id);
+      consoleOutput(`/api/users/${req.params.id}`, "DELETE", 200);
+      return res.status(200).send({ message: `User ${req.params.id} deleted` });
+    }
+    consoleOutput(`/api/users/${req.params.id}`, "DELETE", 404);
+    return res.status(404).send(error[404]);
+  } else {
+    consoleOutput("/api/users", "GET", 403);
+    return res.status(403).send(error[403]);
   }
-  consoleOutput(`/api/users/${req.params.id}`, "DELETE", 404);
-  return res.status(404).send(error[404]);
 };
